# docs/api/postman/whusnet-portal-openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Whusnet Operasional API",
        "version": "0.0.1",
...
    },
    "servers": [
        {
            "url": "https://safari-moon-profession-comedy.trycloudflare.com"
        }
    ],
...
    "paths": {
...
        "/api/customer-portal/auth/refresh": {
            "post": {
                "operationId": "customer-portal.auth.refresh",
                "description": "Menukar refresh token yang masih berlaku dengan pasangan token baru\n(rotasi sekali pakai). Refresh token yang SUDAH dipakai lalu dipakai\nlagi mencabut SEMUA token pelanggan itu \u2014 indikasi pencurian.",
                "summary": "Refresh token",
                "tags": [
                    "Portal Pelanggan"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/PortalRefreshRequest"
                                    },
                                    {
                                        "type": "object",
                                        "properties": {
                                            "refresh_token": {
                                                "type": "string",
                                                "description": "Refresh token yang diterima dari /auth/login atau /auth/refresh sebelumnya.",
                                                "examples": [
                                                    "<64 karakter acak>"
                                                ]
                                            }
                                        },
                                        "required": [
                                            "refresh_token"
                                        ]
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rotasi berhasil, pasangan token baru diterbitkan.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "examples": [
                                        {
                                            "access_token": "<64 karakter acak>",
                                            "refresh_token": "<64 karakter acak>",
                                            "token_type": "Bearer",
                                            "expires_in": 900
                                        }
                                    ],
                                    "properties": {
                                        "access_token": {
                                            "type": "string"
                                        },
                                        "refresh_token": {
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "const": "Bearer"
                                        },
                                        "expires_in": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "access_token",
                                        "refresh_token",
                                        "token_type",
                                        "expires_in"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Sesi tidak valid \u2014 token tidak ditemukan, kedaluwarsa, atau reuse (sudah pernah dipakai sebelumnya).",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "const": "Sesi tidak valid, silakan login ulang."
                                        }
                                    },
                                    "required": [
                                        "message"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                },
                "security": [
                    {
                        "PortalClientSecret": []
                    }
                ]
            }
        },
...
        "/api/customer-portal/auth/logout-all": {
            "post": {
                "operationId": "customer-portal.auth.logout-all",
                "description": "Mencabut SEMUA token pelanggan itu.",
                "summary": "Logout semua perangkat",
                "tags": [
                    "Portal Pelanggan"
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil keluar dari semua perangkat.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "examples": [
                                        {
                                            "message": "Berhasil keluar dari semua perangkat."
                                        }
                                    ],
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "const": "Berhasil keluar dari semua perangkat."
                                        }
                                    },
                                    "required": [
                                        "message"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Access token tidak disertakan atau tidak valid.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "PortalClientSecret": [],
                        "PortalAccessToken": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "PopDistribusiReadToken": {
                "type": "http",
                "description": "Token akses baca-saja untuk membaca struktur topologi Mini POP dan Distribusi.",
                "scheme": "bearer"
            },
            "NetworkAssignmentWriteToken": {
                "type": "http",
                "description": "Token akses tulis untuk mengonfirmasi penugasan jaringan pelanggan.",
                "scheme": "bearer"
            },
            "PortalClientSecret": {
                "type": "apiKey",
                "description": "Client secret portal, statis per environment \u2014 membuktikan \"ini portal resmi\". Dicek di SEMUA endpoint /api/customer-portal/auth/* dan /me/*, TIDAK di /ping.",
                "in": "header",
                "name": "X-Portal-Client"
            },
            "PortalAccessToken": {
                "type": "http",
                "description": "Access token pelanggan (15 menit, dari /auth/login atau /auth/refresh) \u2014 membuktikan \"ini pelanggan X\". TIDAK dipakai di /auth/login, /auth/claim, /auth/refresh (belum ada identitas pelanggan saat endpoint itu dipanggil).",
                "scheme": "bearer"
            }
        },
        "schemas": {
...
            "PortalRefreshRequest": {
                "type": "object",
                "properties": {
                    "refresh_token": {
                        "type": "string"
                    }
                },
                "required": [
                    "refresh_token"
                ],
                "title": "PortalRefreshRequest"
            },
...
        },
        "responses": {
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            },
...
        }
    }
}